requirements.txt
Flask>=2.3,<4
python-pptx>=0.6.21,<1
